Always generate both sync and async
