+++ docs/docs/infrahubctl/infrahubctl-protocols.mdx
@@ -12,8 +12,7 @@ $ infrahubctl protocols [OPTIONS]
 
 * `--branch TEXT`: Branch of schema to export Python protocols for.
 * `--config-file TEXT`: [env var: INFRAHUBCTL_CONFIG; default: infrahubctl.toml]
-* `--out TEXT`: Path to a file to save the result.
-* `--sync / --no-sync`: Generate Python protocols to use with synchronous SDK client.  [default: no-sync]
+* `--out TEXT`: Path to a file to save the result.  [default: schema_protocols.py]
 * `--install-completion`: Install completion for the current shell.
 * `--show-completion`: Show completion for the current shell, to copy it or customize the installation.
 * `--help`: Show this message and exit.
