+++ python_sdk/infrahub_sdk/ctl/cli_commands.py
@@ -20,6 +20,7 @@ from infrahub_sdk.ctl import config
 from infrahub_sdk.ctl.branch import app as branch_app
 from infrahub_sdk.ctl.check import run as run_check
 from infrahub_sdk.ctl.client import initialize_client, initialize_client_sync
+from infrahub_sdk.ctl.constants import PROTOCOLS_TEMPLATE
 from infrahub_sdk.ctl.exceptions import QueryNotFoundError
 from infrahub_sdk.ctl.generator import run as run_generator
 from infrahub_sdk.ctl.render import list_jinja2_transforms
@@ -301,8 +302,7 @@ def transform(
 def protocols(
     branch: str = typer.Option(None, help="Branch of schema to export Python protocols for."),
     _: str = CONFIG_PARAM,
-    out: str = typer.Option(None, help="Path to a file to save the result."),
-    sync: bool = typer.Option(False, help="Generate Python protocols to use with synchronous SDK client."),
+    out: str = typer.Option("schema_protocols.py", help="Path to a file to save the result."),
 ) -> None:
     """Export Python protocols corresponding to a schema."""
 
@@ -335,7 +335,7 @@ def protocols(
 
         return f"{name}: {attribute_kind_map[kind.lower()]}"
 
-    def _jinja2_filter_render_relationship(value: RelationshipSchema) -> str:
+    def _jinja2_filter_render_relationship(value: RelationshipSchema, sync: bool = False) -> str:
         name = value.name
         cardinality = value.cardinality
 
@@ -365,65 +365,6 @@ def protocols(
     client = initialize_client_sync()
     current_schema = client.schema.all(branch=branch)
 
-    template = """
-# Generated by "infrahubctl protocols"
-
-from __future__ import annotations
-
-from typing import TYPE_CHECKING, Protocol, runtime_checkable, Optional
-
-if TYPE_CHECKING:
-    from datetime import datetime
-
-    from infrahub_sdk.nodes import RelatedNode, RelatedNodeSync, RelationshipManager, RelationshipManagerSync
-
-
-@runtime_checkable
-class CoreNode(Protocol):
-    id: str
-    display_label: Optional[str]
-    hfid: Optional[list[str]]
-    hfid_str: Optional[str]
-
-    def get_kind(self) -> str: ...
-
-
-{% for generic in generics %}
-class {{ generic.namespace + generic.name }}(CoreNode):
-    {% if not generic.attributes|default([]) and not generic.relationships|default([]) %}
-    pass
-    {% endif %}
-    {% for attribute in generic.attributes|default([]) %}
-    {{ attribute | render_attribute }}
-    {% endfor %}
-    {% for relationship in generic.relationships|default([]) %}
-    {{ relationship | render_relationship }}
-    {% endfor %}
-    {% if generic.hierarchical | default(false) %}
-    parent: {{ relationship_types.one }}
-    children: {{ relationship_types.many }}
-    {% endif %}
-{% endfor %}
-
-
-{% for node in nodes %}
-class {{ node.namespace + node.name }}({{ node.inherit_from | join(", ") or "CoreNode" }}):
-    {% if not node.attributes|default([]) and not node.relationships|default([]) %}
-    pass
-    {% endif %}
-    {% for attribute in node.attributes|default([]) %}
-    {{ attribute | render_attribute }}
-    {% endfor %}
-    {% for relationship in node.relationships|default([]) %}
-    {{ relationship | render_relationship }}
-    {% endfor %}
-    {% if node.hierarchical | default(false) %}
-    parent: {{ relationship_types.one }}
-    children: {{ relationship_types.many }}
-    {% endif %}
-{% endfor %}
-"""
-
     generics: dict[str, GenericSchema] = {}
     nodes: dict[str, NodeSchema] = {}
 
@@ -438,20 +379,23 @@ class {{ node.namespace + node.name }}({{ node.inherit_from | join(", ") or "Cor
     jinja2_env.filters["render_attribute"] = _jinja2_filter_render_attribute
     jinja2_env.filters["render_relationship"] = _jinja2_filter_render_relationship
 
-    template = jinja2_env.from_string(template)
+    template = jinja2_env.from_string(PROTOCOLS_TEMPLATE)
     rendered = template.render(
-        generics=_sort_and_filter_models(generics),
-        nodes=_sort_and_filter_models(nodes),
-        relationship_types={
-            "one": "RelatedNodeSync" if sync else "RelatedNode",
-            "many": "RelationshipManagerSync" if sync else "RelationshipManager",
-        },
+        generics=_sort_and_filter_models(generics), nodes=_sort_and_filter_models(nodes), sync=False
+    )
+    rendered_sync = template.render(
+        generics=_sort_and_filter_models(generics), nodes=_sort_and_filter_models(nodes), sync=True
     )
 
+    output_file = Path(out)
+    output_file_sync = Path(output_file.stem + "_sync" + output_file.suffix)
+
     if out:
-        write_to_file(Path(out), rendered)
+        write_to_file(output_file, rendered)
+        write_to_file(output_file_sync, rendered_sync)
     else:
         console.print(rendered)
+        console.print(rendered_sync)
 
 
 @app.command(name="version")
